### 2020-11-02 #406 @CarlosRealR
> Hello @CarlosRealR, thanks for submitting. We will assign you a handling editor soon. In the meantime, can you please elaborate on the answers to the questions of the Scope section? (in the opening post)

Hello @melvidoni, done. I hope my answers would be clear enough now.
