### 2020-11-03 #406 @ldecicco-USGS
Hi @CarlosRealR , I'll begin asking around for 2 reviewers and running the automated package checks. Thanks for expanding on the questions above 👍

One thing that would be helpful while I'm asking around for reviewers would be to "Knit" the Readme.Rmd, and then check-in the Readme.md. The main difference is that Github will then render the file. You _could_ copy a few of the examples in the vignette to showcase the main idea of the package in a pretty/rendered way. So, not required, I'm just thinking it might help the potential reviewers decide if they are a good fit for taking on the review.

Also, not sure how responsive my American colleagues are going to be this week. I'd be prepared for a slightly slower response for the next couple of days than maybe usual.
