### 2020-11-03 #406 @CarlosRealR
Hi @ldecicco-USGS , excuse me, but this is the first time I make a package and send software to GitHub and I'm a a bit confused with the procedure and the best way to do things. What do you mean with "check-in the Readme.md"? I would be grateful if you could tell me a package with a readme that way to understand what I have to do exactly.
