### 2020-11-03 #406 @ldecicco-USGS
Once you knit the Rmd file, it will create an md file, so now on your computer there will be 2 files, the "Rmd" one is specific to the "R" programming language, the "md" is a markdown file (more universal). So not only do you need to generate the "md" file, you need to get it up to Github. If you are in RStudio using the "Git" tab, that will mean clicking on the checkbox next to the new file, hitting "commit", writing a little message, then clicking the "Push" button.

Probably "check-in" was the wrong term, sorry! If you are doing it in the Terminal, it would be something like:
```
git add README.md
git commit  -m "Adding the md file"
git push origin master
```
