### 2020-11-04 #406 @CarlosRealR
Hi, I expanded the README.rd including some images as you suggested. Thanks for the advice.
