### 2020-11-13 #406 @ldecicco-USGS
Just wanted to check in and let you know I'm sending out some emails today. I ran the automatic checks and everything looks good:

```r
goodpractice::gp()
Preparing: covr
Preparing: cyclocomp
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck

♥ Yee-haw! Well-made package! Keep up the prime work!
```

```r
spelling::spell_check_package()
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD            FOUND IN
barycenter      pixelclasser.Rmd:85
carlos          pixelclasser.Rd:28
colour          classify_pixels.Rd:16,28
                define_cat.Rd:12,24
                define_rule.Rd:12,15,34,36,48
                pixelclasser.Rd:6,9,11
                plot_pixels.Rd:12,14,16,27,29,30,31,35
                plot_rgb_plane.Rd:17,19,38
                plot_rule.Rd:16,34,37
                read_image.Rd:18
                description:2
                README.md:11,16,67
                README.Rmd:25,27,54
                pixelclasser.Rmd:21,23,25,62,78,85,114,151,184
Colour          title:1
coloured        pixelclasser.Rmd:85
colours         classify_pixels.Rd:10
                define_cat.Rd:47
                plot_pixels.Rd:32
                read_image.Rd:16
                pixelclasser.Rmd:21,66,96
comparator      define_rule.Rd:63
                pixelclasser.Rmd:132
correctedness   pixelclasser.Rmd:121
gb              pixelclasser.Rmd:96
geq             pixelclasser.Rmd:25
incid           classify_pixels.Rd:41
jpeg            pixelclasser.Rd:9
                read_image.Rd:41
                save_classif_image.Rd:26
                description:2
jpg             read_image.Rd:5
                save_classif_image.Rd:30
ldots           pixelclasser.Rmd:27,29
leq             pixelclasser.Rmd:25
mathbf          pixelclasser.Rmd:27,27,29,29
migth           pixelclasser.Rmd:76
mis             pixelclasser.Rmd:161,254
pertenence      README.md:16
                README.Rmd:27
                pixelclasser.Rmd:23
poligons        pixelclasser.Rmd:172
rgb             plot_pixels.Rd:23
                pixelclasser.Rmd:23,62,64,76
subcat          define_cat.Rd:17
subclasses      pixelclasser.Rmd:184,190
usc             pixelclasser.Rd:28
```
Some of those are not actually mis-spelled, but you might want to check a few.

Please consider adding an rOpenSci review badge -- either via -- `rodev::use_review_badge(406)`.
