### 2020-12-01 #406 @qdread
@ldecicco-USGS and @CarlosRealR, here is my package review! This is my first review for ROpenSci, so please be understanding if I've left anything out or otherwise not followed best reviewer practices. Thanks for the opportunity.

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

(left out the section on JOSS)

---

This package is essentially a way to walk the user through the process of manually classifying pixels in an image into different categories based on their relative proportions of red, blue, and green. The user does this by plotting the pixels in 2 of the 3 color dimensions, eyeballing the dividing lines between groups of pixels belonging to differently categorized regions of the image, specifying the coordinates of the line, and then repeating until the cloud of points is "tiled" into convex polygons, each of which is assigned to a category. The resulting classification rules could then be applied to other images. Overall, I found that the package generally follows the ROpenSci package guidelines, and that it works as advertised. Nice job! I do have some general/conceptual comments, as well as some technical comments about the code and documentation.

- My main general comment on the statement on functionality is that it seems slightly exaggerated to call this "a simple version of the multivariate technique known as Support Vector Machine." Support Vector Machine is a technique that finds the optimal hyperplane to separate classes in a dataset. Furthermore, SVM is a general and flexible method that includes, for example, nonlinear boundaries between classes. So, I think the comparison is not quite apt. I think a prospective user might expect that there would be some way of identifying the "best" boundary between classes, rather than relying entirely on manual separation. This isn't necessarily a problem with the functionality itself, just the way it is described.
- One other general suggestion is that if the rules are defined manually, it is sometimes slow and awkward to do this by trial and error, iteratively changing the coordinates of the two endpoints of the rule and then plotting the rule. The function `graphics::locator()` can help with this. It might be worthwhile to include a wrapper function for `locator()` with some preset options that the typical user of this package might select.
- The order in which the functions need to be run should be more clear from the help documentation (not just the vignette). This is because the functions need to be run in a very specific sequence so the principle of ["multiple points of entry"](https://devguide.ropensci.org/building.html#general) is broken a little &mdash; you can't tell from some of the help documentation where in the process you are.
- I think that all the classes of objects you define should have a `print` method so that informative output is displayed when a user enters an object's name into the console. It could be as simple as displaying (abbreviated) output of `str(object_name)` when `object_name` is called. Currently, calling the name of some of the objects returns an error (see below under `define_rule`).
- My final comment is that I tried this method out on some other images (some of leaves and some of cells) that I found on a Google Images search. It did not work well on the pictures of leaves because the contrast between live and dead leaves was not as obvious as in the example image provided with the package. The cell image worked well because the cell structures were stained so that the contrast between classes was extremely high. Generally I think the package is nice and simple, which is a good thing, but I think that users will find that it will only work effectively for images that have a high and obvious contrast between classes.

#### Vignette

- The vignette introduction could be better laid out, with bullet points, etc., and refer to relevant literature or other resources for further reading on the theory underlying the methods.
- I was very confused at first why rule 3 and rule 5 had different coordinates, although they were intended to be the same line. I see that you intentionally introduced an error to illustrate how you might find and correct a typo, but I found it quite confusing. I would suggest omitting the typo.

#### Comments on specific functions

`define_rule`

- I don't think that "comparator" is a widely understood word so it might not be the best choice for an argument name. Maybe just use "comparison" instead?
- As mmentioned above in the general comments, the `rule` objects should really have a print method. Perhaps just show the output of `str(rule)` when `rule` is entered in the console. Currently, I get the following error when I call the name of a `rule` object (which also appears for `pixel_cat` objects):

```
Error in cat(x, ..., sep = sep) :
  argument 1 (type 'list') cannot be handled by 'cat'
```

- In the output of `define_rule`, the elements `a` and `c` are not very informative names. Maybe just call them `slope` and `intercept`? I have not often seen the notation `ax + c` for a regression line. More commonly the slope and intercept are called `a` and `b`, or `m` and `b`, or `beta` notation is used.

`plot_rgb_plane`

- This function works well but what I really would have liked is a function that plots all three combinations of two colors in three separate panels. Then, `plot_pixels` could also be rewritten to add points to the three panels, and `plot_rule` could be rewritten to only add the rule to the appropriate panel.

`plot_rule`

- It is slightly confusing that some of the graphical parameter arguments passed with `...` are arguments to `lines()` and others to `text()`. Is there a good way to separate those, perhaps with two lists of arguments? That would also enable, for example, different colors for the line and the text label if two separate lists of arguments were passed. Also, the help documentation only states that `...` are passed to `lines()` although the arguments are also passed to `text()`.

`plot_pixels`

- It would be an improvement to be able to pass graphical parameters to `points()` via a `...` argument, because `plot_pixels` is just a wrapper for `points`.

`classify_pixels`

- I think you should spell out the word "category" in the output rather than using the nonstandard abbreviation "categ."
- It would be helpful if the summary totals of pixel categories could be printed as part of the object's `print` method, rather than only when the classification is run.
- I am not sure if "redundant" is the appropriate term for pixels classified into >1 category. To me, a "redundant" feature is one that does not add information to the classification (for example colinear with another feature), rather than a data point that is classified in two categories.

#### Comments on package documentation

- The `testthat` tests cover all the exported functions adequately.
- You should add a newline to the bottom of the `DESCRIPTION` file to avoid the warning that the final line is incomplete.
- Note on community guidelines: I did not see any contribution guidelines, so it would be helpful to include those. The `DESCRIPTION` does not include a `Maintainer` field (although Carlos Real is listed as `cre` under `Authors@R`).
- Minor note on examples: The example for `save_classif_image()` does not run locally but I think that is fine in this case because it would be too cumbersome to include a working example for that function.
