### 2020-12-01 #406 @ldecicco-USGS
Thanks @qdread ! Looks great to me. @CarlosRealR feel free to respond to @qdread 's comments or wait for @superjai
