### 2020-12-02 #406 @CarlosRealR
Thanks @qdread  for your favourable opinion and the helpful comments. I am making some of the suggested changes, but I will wait for the comments of the other reviewer to make the bulk of the changes and to send a joint response. I think it is the best way to proceed.
