### 2021-01-07 #406 @CarlosRealR
Hi @ldecicco-USGS,
I am asking the editor of Methods in Ecology and Evolution for an extension of the deadline for resubmitting the paper about  pixelclasser, as the one they gave us (january 11) is about to expire. I know that the end of a year are busy times for everybody but, do you know approximately when will we receive the comments of the second reviewer? I am considering whether I must negotiate for a 15 day delay or better for a month. I suppose that MEE will not reject my request, but I prefer asking for too much time now than to ask for successive extensions, less work for everybody.
