### 2021-01-15 #406 @ldecicco-USGS
I'm submitting a review due to how long this has taken. If the other reviewer adds more thoughts in the future, you can take that into account in your long-term maintenance and development of the package, but for now just be concerned with this review and qdread's above.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a nice package that performs some very useful functions, and includes ways for the user to have fine control over the way they classify images.

The examples all ran on my computer except the ones that explicitly stated they were hypothetical. There were several examples wrapped in `dontrun` that I'm not sure why you would not want them run. They were relatively fast and useful. I like to keep as many examples and tests in my checks as possible for long-term stability and sustainability.

I ran the workflow described in the vignette on an image I had on my computer. It was exceptionally slow, but eventually things finished. As I walked through the example with my own image, I could see how useful the package is.

All-in-all, handy package. The code itself looked straightforward and did not raise any red flags.

FYI, my `goodpractice::gp()` output was:
```
♥ Hooray! Super-excellent package! Keep up the classy work!
```
