### 2021-01-15 #406 @CarlosRealR
Thanks for your comments @ldecicco-USGS. I have done most of the modifications suggested by the first reviewer yet, but changes to tests and the vignette were waiting. I will send them on Monday or Thursday. May be that the comments of the second reviewer arrive before that.
