### 2021-01-19 #406 @CarlosRealR
Thanks for the helpful comments about the package, @qdread and @ldecicco-USGS .

Before explaining the changes I wish to make some general comments.

The method is true SVM. In papers explaining the technique (Cortes and Vapnik, 1995; Bennet and Campbell, 2000, now cited in the vignette), the examples illustrating the basics of SVM are very similar to what pixelclasser does (they are 2D examples). Linear regression is the simplest technique of the family of regression methods (including multivariate regression) but it is regression nevertheless, even when you only fits the line without making significance tests. With pixelclasser you can do SVM in a manual, simple way, but it remains (bare bones) SVM.

It is true that the method cannot discriminate categories that are not separable (overlapping). In fact, the example in the vignette was selected to show this: first a clean classification and then a dirty one. Therefore, it is not a universal method for image classification, but it will work in many cases and with much less complexity than other methods that are in use. It also gives the user full control and understanding of what is going on, something that complex libraries and canned code don’t do in many cases.

Processing the data can be slow, specially when plotting the pixels. Note that many cameras can produce 20 million points per image (and 100 is not unusual in hi-res cameras). R plotting routines choke with such a lot of data. I do not see an easy solution to this, (apart from using a sample of the pixels, which distorts the cloud of points). Using a fast computer with lots of memory helps, as usual. I have added a warning in the vignette and in the help page of plot_pixels().

These are the changes and additions:
    • An outline of the workflow was included in the help page of the package (?pixelclasser).
    • New functions:
        ◦ place_rule() wraps a call to graphics::locator() and helps in drawing vertical and horizontal lines. define_rule() can be used as in the previous version and also can be combined with place_rule() to create rules easily. This is illustrated in the vignette and in the examples of its help page.
        ◦ Labelling the rules is now done with an independent function, label_rule(). In this way label colours, for example, can be set independently from the line colour.
    • The names of the objects are more coherent now, all starting by pixel_
    • All objects have now its print method.
        ◦ The print method for the pixel_rule objects writes the equation of the line using the colour variables, and without using generic parameters like a and c).
        ◦ “redundant” has been substituted by “duplicate” in the output of classify_pixels() .
    • “comparator” has been substituted by “comp.op”
    • plot_pixels() has now a … argument to pass parameters to the underlying points() function.
    • The layout of the vignette introduction was improved (I hope).
    • I left the error in the definition of rule_05 to show later the usefulness of the checks in classify_pixels(), but flagged and explained it on the spot to avoid confusing the reader.

And this is what was not done:

Due to the problem with plotting a lot of points commented above, trying to plot the three axis combinations in the same graph is not a good idea, so I decided to left the plot functions unchanged.

I think that I addressed all the questions raised in the reviews. I have pushed version 0.7 to the repository. It contains all these changes.
