### 2021-01-19 #406 @qdread
@CarlosRealR Wow this looks great! I'm glad my comments were able to improve your package. @ldecicco-USGS am I expected to do a second review?
