### 2021-01-20 #406 @qdread
Hi @CarlosRealR I just went over the vignette of your revised package and I must say again, great job! 💪 🙇 You really went above and beyond in addressing my comments. I especially appreciate how the `print` methods now show concise and informative details on each object, and I like the `place_rule()` and `label_rule()` functions! I ran the package checks and everything was 💯 .

I just wanted to note a few small things I noticed:

(1) I got this warning on install:

```
Rd warning: C:/Users/qread/AppData/Local/Temp/RtmpO0Jg7d/R.INSTALL462c7533769d/pixelclasser/man/classify_pixels.Rd:72: file link 'colours' in package 'grDevices' does not exist and so has been treated as a topic
```

Maybe this is because of a difference between `colors` and `colours`? Not sure.

(2) I got this error when running the code chunk in the vignette demonstrating the `place_rule()` function:

```
Error in define_rule("rule_07", "g", "b", rp02, ">=") :
  rule_points must be a list containing two coordinate vectors
```

That error happened even though `rp02` was an object of class `rule_points` that I created using `rp02 <- place_rule("v")`.

(3) I didn't see contribution guidelines in the documentation and I didn't know if you wanted to include those.

Anyway, I wanted to bring those small things to your attention. But again, great work and I think my job here is done 😁
