### 2021-01-21 #406 @CarlosRealR
Hi @qdread . I am now correcting the errors that you found in define_rule. I was not able to reproduce the install warning in my linux machine but I will try in a Windows one to see if it appears. Please, could you explain me what are these "contribution guidelines" that should be in the documentation. I have searched in "Writing R extensions" and I found nothing. Is this an ROpenSci issue? I am lost here...
