### 2021-01-21 #406 @qdread
Hi @CarlosRealR I don't know if it is a big deal to include contribution guidelines but as you can see in the review checklist under "documentation" there is a checkbox for "community guidelines" potentially including contribution guidelines. Many packages include a file called `CONTRIBUTING.md` that explains how to contribute to the package. It might be as simple as saying open a pull request on the GitHub repo for the package.

Here are a couple examples

- https://corybrunson.github.io/ggalluvial/CONTRIBUTING.html
- https://www.tidyverse.org/blog/2017/08/contributing/

But I don't think it needs to be too complex. I will leave it up to the ROpenSci people to say whether this is something that absolutely needs to be included or not.
