### 2021-01-25 #406 @CarlosRealR
Hi . @ldecicco-USGS and @qdread. I have just pushed the revised version of pixelclasser. I have corrected the errors that @qdread  find in the previous version, I finally tracked them all. I changed a little the interface of place_label() and polished the whole set of function help files. In these, I corrected some inconsistencies and some crap from old versions that managed to escape my previous inspection. I also modified some examples which, as @ldecicco-USGS commented, can run without problem and were previously marked as not to run.
I also added a very simple CONTRIBUTION file, simply recommending to send an e-mail to me. When the package will be finally accepted, will be the time to add gitHub issues and the like, I think.
