### 2021-01-25 #406 @ldecicco-USGS
Approved! Thanks @CarlosRealR for submitting and @qdread for your review! 😸

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with Travis CI and GitHub Actions)
- [ ] If you plan to submit to CRAN, check out our [CRAN tips](https://devguide.ropensci.org/building.html#crangotchas)
