### 2021-01-26 #406 @CarlosRealR
Good news! Thanks again for the comments  on the package, @ldecicco-USGS  and @qdread. They helped to improve it, really. Now I am doing a last review before convert it to the version 1.0. I am including @qdread in the credits as reviewer, but I think that you, @ldecicco-USGS has also done review work here. If you wish, I will include you as reviewer, but perhaps this is not compatible with the role of editor, or redundant. Tell me if this convenient or not, please.
