### 2021-01-26 #406 @CarlosRealR
I have a question @ldecicco-USGS
I tried to transfer the repo as indicated in your comment. As I understand this, I must use the option Transfer ownership in the Danger Zone of the Settings of my repository, indicating ropensci as the organization owning the new repo. But I receive this message:
 ropensci/pixelclasser already exists and You don’t have the permission to create public repositories on ropensci

and nothing happens.

I have also tried the git instructions that appear in ropensci/pixelclasser, but these do not work neither. I think that my configuration points to my repo, so I must dig a bit in how to change my configuration. But I am not sure if I am correct or not. Could you give me some tip about this?
