### 2021-02-10 #406 @ldecicco-USGS
@CarlosRealR It looks like `pixelclasser` is all set. I'm going to close the issue, but feel free to re-open if there are lingering issues/questions. Thanks again for the submission!
